Remove the before in cli to support --clean
I was using the Before func to prep the env but that disconnected the
resulting environment from actual command. Rather than figuring out
how to make that work within the context of cli, I moved the calling
of the command to a single action function.

I have a feeling I could remove the cli dep in favor of something else
as well like pflags.