Refactor to support templates in alias files
This makes a subtle change where the template is not automatically
rendered at the end of a run. I think this is OK b/c it is easy enough
for a user to add any template calls last as a best practice.

There is also the issue of how to clean things up afterwards, but
again, that can be a different flag that can show up in the alias /
base parser.